move auth function tests
