+++ api-js/src/auth/__tests__/check-domain-ownership.test.js
@@ -3,12 +3,12 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { setupI18n } = require('@lingui/core')
 
-const { makeMigrations } = require('../../migrations')
-const { checkDomainPermission } = require('../auth')
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
+const { makeMigrations } = require('../../../migrations')
+const { checkDomainOwnership } = require('..')
+const englishMessages = require('../../locale/en/messages')
+const frenchMessages = require('../../locale/fr/messages')
 
-describe('given the check domain permission function', () => {
+describe('given the check domain ownership function', () => {
   let query, drop, truncate, migrate, collections, org, domain, i18n
 
   let consoleOutput = []
@@ -60,7 +60,7 @@ describe('given the check domain permission function', () => {
       lastRan: null,
       selectors: ['selector1', 'selector2'],
     })
-    await collections.claims.save({
+    await collections.ownership.save({
       _to: domain._id,
       _from: org._id,
     })
@@ -71,7 +71,7 @@ describe('given the check domain permission function', () => {
     await drop()
   })
 
-  describe('given a successful domain permission check', () => {
+  describe('given a successful domain ownership check', () => {
     let user, permitted
     beforeEach(async () => {
       const userCursor = await query`
@@ -81,7 +81,7 @@ describe('given the check domain permission function', () => {
       `
       user = await userCursor.next()
     })
-    describe('if the user belongs to an org which has a claim for a given organization', () => {
+    describe('if the user belongs to an org which has a ownership for a given organization', () => {
       afterEach(async () => {
         await query`
           LET userEdges = (FOR v, e IN 1..1 ANY ${org._id} affiliations RETURN { edgeKey: e._key, userId: e._to })
@@ -102,11 +102,13 @@ describe('given the check domain permission function', () => {
           })
         })
         it('will return true', async () => {
-          const testCheckDomainPermission = checkDomainPermission({
+          const testCheckDomainOwnerShip = checkDomainOwnership({
             query,
             userId: user._key,
           })
-          permitted = await testCheckDomainPermission({ domainId: domain._id })
+          permitted = await testCheckDomainOwnerShip({
+            domainId: domain._id,
+          })
           expect(permitted).toEqual(true)
         })
       })
@@ -119,11 +121,13 @@ describe('given the check domain permission function', () => {
           })
         })
         it('will return true', async () => {
-          const testCheckDomainPermission = checkDomainPermission({
+          const testCheckDomainOwnerShip = checkDomainOwnership({
             query,
             userId: user._key,
           })
-          permitted = await testCheckDomainPermission({ domainId: domain._id })
+          permitted = await testCheckDomainOwnerShip({
+            domainId: domain._id,
+          })
           expect(permitted).toEqual(true)
         })
       })
@@ -136,18 +140,20 @@ describe('given the check domain permission function', () => {
           })
         })
         it('will return true', async () => {
-          const testCheckDomainPermission = checkDomainPermission({
+          const testCheckDomainOwnerShip = checkDomainOwnership({
             query,
             userId: user._key,
           })
-          permitted = await testCheckDomainPermission({ domainId: domain._id })
+          permitted = await testCheckDomainOwnerShip({
+            domainId: domain._id,
+          })
           expect(permitted).toEqual(true)
         })
       })
     })
   })
 
-  describe('given an unsuccessful domain permission check', () => {
+  describe('given an unsuccessful domain ownership check', () => {
     let user
     beforeEach(async () => {
       const userCursor = await query`
@@ -157,18 +163,20 @@ describe('given the check domain permission function', () => {
       `
       user = await userCursor.next()
     })
-    describe('if the user does not belong to an org which has a claim for a given organization', () => {
+    describe('if the user does not belong to an org which has a ownership for a given domain', () => {
       let permitted
       it('will return false', async () => {
-        const testCheckDomainPermission = checkDomainPermission({
+        const testCheckDomainOwnerShip = checkDomainOwnership({
           query,
           userId: user._key,
         })
-        permitted = await testCheckDomainPermission({ domainId: domain._id })
+        permitted = await testCheckDomainOwnerShip({
+          domainId: domain._id,
+        })
         expect(permitted).toEqual(false)
       })
     })
-    describe('language is set to english', () => {
+    describe('users language is set to english', () => {
       beforeAll(() => {
         i18n = setupI18n({
           language: 'en',
@@ -180,30 +188,34 @@ describe('given the check domain permission function', () => {
           },
         })
       })
-      describe('if a database error is encountered during permission check', () => {
+      describe('if a database error is encountered during ownership check', () => {
         let mockQuery
         it('returns an appropriate error message', async () => {
           mockQuery = jest
             .fn()
             .mockRejectedValue(new Error('Database error occurred.'))
           try {
-            const testCheckDomainPermission = checkDomainPermission({
+            const testCheckDomainOwnerShip = checkDomainOwnership({
               i18n,
               query: mockQuery,
               userId: user._key,
             })
-            await testCheckDomainPermission({ domainId: domain._id })
+            await testCheckDomainOwnerShip({
+              domainId: domain._id,
+            })
           } catch (err) {
             expect(err).toEqual(
-              new Error('Authentication error. Please sign in again.'),
+              new Error(
+                'Error when retrieving dmarc report information. Please try again.',
+              ),
             )
             expect(consoleOutput).toEqual([
-              `Error when retrieving affiliated organization claims for user with ID ${user._id} and domain with ID ${domain._id}: Error: Database error occurred.`,
+              `Database error when retrieving affiliated organization ownership for user: ${user._id} and the domain: ${domain._id}: Error: Database error occurred.`,
             ])
           }
         })
       })
-      describe('if a cursor error is encountered during permission check', () => {
+      describe('if a cursor error is encountered during ownership check', () => {
         let mockQuery
         it('returns an appropriate error message', async () => {
           const cursor = {
@@ -213,24 +225,28 @@ describe('given the check domain permission function', () => {
           }
           mockQuery = jest.fn().mockReturnValue(cursor)
           try {
-            const testCheckDomainPermission = checkDomainPermission({
+            const testCheckDomainOwnerShip = checkDomainOwnership({
               i18n,
               query: mockQuery,
               userId: user._key,
             })
-            await testCheckDomainPermission({ domainId: domain._id })
+            await testCheckDomainOwnerShip({
+              domainId: domain._id,
+            })
           } catch (err) {
             expect(err).toEqual(
-              new Error('Unable to find domain. Please try again.'),
+              new Error(
+                'Error when retrieving dmarc report information. Please try again.',
+              ),
             )
             expect(consoleOutput).toEqual([
-              `Error when retrieving affiliated organization claims for user with ID ${user._id} and domain with ID ${domain._id}: Error: Cursor error occurred.`,
+              `Cursor error when retrieving affiliated organization ownership for user: ${user._id} and the domain: ${domain._id}: Error: Cursor error occurred.`,
             ])
           }
         })
       })
     })
-    describe('language is set to french', () => {
+    describe('users language is set to french', () => {
       beforeAll(() => {
         i18n = setupI18n({
           language: 'fr',
@@ -242,28 +258,30 @@ describe('given the check domain permission function', () => {
           },
         })
       })
-      describe('if a database error is encountered during permission check', () => {
+      describe('if a database error is encountered during ownership check', () => {
         let mockQuery
         it('returns an appropriate error message', async () => {
           mockQuery = jest
             .fn()
             .mockRejectedValue(new Error('Database error occurred.'))
           try {
-            const testCheckDomainPermission = checkDomainPermission({
+            const testCheckDomainOwnerShip = checkDomainOwnership({
               i18n,
               query: mockQuery,
               userId: user._key,
             })
-            await testCheckDomainPermission({ domainId: domain._id })
+            await testCheckDomainOwnerShip({
+              domainId: domain._id,
+            })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
             expect(consoleOutput).toEqual([
-              `Error when retrieving affiliated organization claims for user with ID ${user._id} and domain with ID ${domain._id}: Error: Database error occurred.`,
+              `Database error when retrieving affiliated organization ownership for user: ${user._id} and the domain: ${domain._id}: Error: Database error occurred.`,
             ])
           }
         })
       })
-      describe('if a cursor error is encountered during permission check', () => {
+      describe('if a cursor error is encountered during ownership check', () => {
         let mockQuery
         it('returns an appropriate error message', async () => {
           const cursor = {
@@ -273,16 +291,18 @@ describe('given the check domain permission function', () => {
           }
           mockQuery = jest.fn().mockReturnValue(cursor)
           try {
-            const testCheckDomainPermission = checkDomainPermission({
+            const testCheckDomainOwnerShip = checkDomainOwnership({
               i18n,
               query: mockQuery,
               userId: user._key,
             })
-            await testCheckDomainPermission({ domainId: domain._id })
+            await testCheckDomainOwnerShip({
+              domainId: domain._id,
+            })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
             expect(consoleOutput).toEqual([
-              `Error when retrieving affiliated organization claims for user with ID ${user._id} and domain with ID ${domain._id}: Error: Cursor error occurred.`,
+              `Cursor error when retrieving affiliated organization ownership for user: ${user._id} and the domain: ${domain._id}: Error: Cursor error occurred.`,
             ])
           }
         })
