+++ api-js/src/auth/__tests__/check-permission.test.js
@@ -3,10 +3,10 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { setupI18n } = require('@lingui/core')
 
-const { makeMigrations } = require('../../migrations')
-const { checkPermission } = require('../auth')
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
+const { makeMigrations } = require('../../../migrations')
+const { checkPermission } = require('..')
+const englishMessages = require('../../locale/en/messages')
+const frenchMessages = require('../../locale/fr/messages')
 
 describe('given the check permission function', () => {
   let query, drop, truncate, migrate, collections, i18n
