+++ api-js/src/auth/__tests__/user-required.test.js
@@ -3,11 +3,11 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { setupI18n } = require('@lingui/core')
 
-const { makeMigrations } = require('../../migrations')
-const { userLoaderByKey } = require('../loaders')
-const { userRequired } = require('../auth')
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
+const { makeMigrations } = require('../../../migrations')
+const { userLoaderByKey } = require('../../loaders')
+const { userRequired } = require('..')
+const englishMessages = require('../../locale/en/messages')
+const frenchMessages = require('../../locale/fr/messages')
 
 describe('given a userLoaderByKey dataloader', () => {
   let query, drop, truncate, migrate, collections, i18n
